+++ contracts/contract/Governance.sol
@@ -10,7 +10,7 @@ import "@openzeppelin/contracts/security/ReentrancyGuard.sol";
 import "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
 
 contract Governance is IGovernance, ReentrancyGuard {
-    uint256 private immutable version;
+    uint8 private immutable version;
     uint256 private immutable thresholdVotingPower;
 
     bytes32 public validatorSetHash;
@@ -23,7 +23,7 @@ contract Governance is IGovernance, ReentrancyGuard {
     IHub private hub;
 
     constructor(
-        uint256 _version,
+        uint8 _version,
         address[] memory _validators,
         uint256[] memory _powers,
         uint256 _thresholdVotingPower,
